move sharing client to separate file, update refrences, fix tests
